refactor: move logger to separate file
