+++ src/utils/index.ts
@@ -79,39 +79,3 @@ export const createMultipleEditButton = (
 
     return multipleDiv;
 };
-
-export class PublicLogger {
-    private static logEvent(
-        logCallback:
-            | Console["log"]
-            | Console["warn"]
-            | Console["error"]
-            | Console["debug"]
-            | Console["info"],
-        message: any[]
-    ): void {
-        if (process.env.NODE_ENV !== "test") {
-            logCallback("Live_Preview_SDK:", ...message);
-        }
-    }
-
-    public static log(...data: any[]): void {
-        this.logEvent(console.log, data);
-    }
-
-    public static error(...data: any[]): void {
-        this.logEvent(console.error, data);
-    }
-
-    public static warn(...data: any[]): void {
-        this.logEvent(console.warn, data);
-    }
-
-    public static info(...data: any[]): void {
-        this.logEvent(console.info, data);
-    }
-
-    public static debug(...data: any[]): void {
-        this.logEvent(console.debug, data);
-    }
-}
